requirements.txt
streamlit
langchain-community
langchain-huggingface
langchain-google-genai
python-dotenv
pypdf
docx2txt
faiss-cpu
youtube-transcript-api